+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Unreleased
 
 * Drop cache TTL to 60 seconds.
+* Remove the component system, components are now consumed via the [govuk_publishing_components gem](https://github.com/alphagov/govuk_publishing_components)
 
 # 12.1.0